+++ CHANGELOG.md
@@ -75,6 +75,7 @@ Putting this stuff down here so things don't conflict as often.
 * Fixed chest item changes not triggering any range checks, tile checks, or correct chest checks. (@hakusaro)
 * Added `TSPlayer` to `GetDataHandlers.PlayerBuff`. (@hakusaro)
 * Added `TSPlayer` and `PlayerDeathReason` to `GetDataHandlers.PlayerDamage`. (@hakusaro)
+* Added `TSPlayer` to `GetDataHandlers.NPCStrike`. (@hakusaro)
 
 ## TShock 4.3.25
 * Fixed a critical exploit in the Terraria protocol that could cause massive unpreventable world corruption as well as a number of other problems. Thanks to @bartico6 for reporting. Fixed by the efforts of @QuiCM, @hakusaro, and tips in the right directioon from @bartico6.